Add support for checking minimal permissions for any API without APIC. Closes #861
