+++ dev-proxy-plugins/MinimalPermissions/ApiPermissionError.cs
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace Microsoft.DevProxy.Plugins.MinimalPermissions;
+
+public class ApiPermissionError
+{
+    public required string Request { get; init; }
+    public required string Error { get; init; }
+}
