+++ dev-proxy-plugins/MinimalPermissions/GraphPermissionError.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Text.Json.Serialization;
+
+namespace Microsoft.DevProxy.Plugins.MinimalPermissions;
+
+internal class GraphPermissionError
+{
+    [JsonPropertyName("requestUrl")]
+    public string Url { get; set; } = string.Empty;
+    public string Message { get; set; } = string.Empty;
+}
