+++ dev-proxy-plugins/MinimalPermissions/GraphPermissionsType.cs
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace Microsoft.DevProxy.Plugins.MinimalPermissions;
+
+public enum GraphPermissionsType
+{
+    Application,
+    Delegated
+}
