+++ dev-proxy-plugins/MinimalPermissions/GraphResultsAndErrors.cs
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace Microsoft.DevProxy.Plugins.MinimalPermissions;
+
+internal class GraphResultsAndErrors
+{
+    public GraphPermissionInfo[]? Results { get; set; }
+    public GraphPermissionError[]? Errors { get; set; }
+}
