--- dev-proxy-plugins/MinimalPermissions/PermissionError.cs
@@ -1,10 +1,0 @@
-using System.Text.Json.Serialization;
-
-namespace Microsoft.DevProxy.Plugins.MinimalPermissions;
-
-internal class PermissionError
-{
-    [JsonPropertyName("requestUrl")]
-    public string Url { get; set; } = string.Empty;
-    public string Message { get; set; } = string.Empty;
-}
