--- dev-proxy-plugins/MinimalPermissions/PermissionsType.cs
@@ -1,7 +1,0 @@
-namespace Microsoft.DevProxy.Plugins.MinimalPermissions;
-
-public enum PermissionsType
-{
-    Application,
-    Delegated
-}
