--- dev-proxy-plugins/MinimalPermissions/RequestInfo.cs
@@ -1,11 +1,0 @@
-
-using System.Text.Json.Serialization;
-
-namespace Microsoft.DevProxy.Plugins.MinimalPermissions;
-
-public class RequestInfo
-{
-    [JsonPropertyName("requestUrl")]
-    public string Url { get; set; } = string.Empty;
-    public string Method { get; set; } = string.Empty;
-}
