--- dev-proxy-plugins/MinimalPermissions/ResultsAndErrors.cs
@@ -1,7 +1,0 @@
-namespace Microsoft.DevProxy.Plugins.MinimalPermissions;
-
-internal class ResultsAndErrors
-{
-    public PermissionInfo[]? Results { get; set; }
-    public PermissionError[]? Errors { get; set; }
-}
