+++ dev-proxy-plugins/OpenApi/OpenApiDocumentExtensions.cs
@@ -2,11 +2,14 @@
 // Licensed under the MIT License.
 
 using System.Text.RegularExpressions;
+using Microsoft.DevProxy.Abstractions;
+using Microsoft.DevProxy.Plugins.MinimalPermissions;
 using Microsoft.Extensions.Logging;
 
 #pragma warning disable IDE0130
 namespace Microsoft.OpenApi.Models;
 #pragma warning restore IDE0130
+
 
 public static class OpenApiDocumentExtensions
 {
@@ -113,5 +116,102 @@ public static class OpenApiDocumentExtensions
             .Where(s => s.Value.Type == SecuritySchemeType.OAuth2)
             .Select(s => s.Value)
             .ToArray();
+    }
+
+    public static ApiPermissionsInfo CheckMinimalPermissions(this OpenApiDocument openApiDocument, IEnumerable<RequestLog> requests, ILogger logger)
+    {
+        logger.LogInformation("Checking minimal permissions for API {apiName}...", openApiDocument.Servers.First().Url);
+
+        var tokenPermissions = new List<string>();
+        var operationsFromRequests = new List<ApiOperation>();
+        var operationsAndScopes = new Dictionary<string, string[]>();
+        var errors = new List<ApiPermissionError>();
+
+        foreach (var request in requests)
+        {
+            // get scopes from the token
+            var methodAndUrl = request.MessageLines.First();
+            var methodAndUrlChunks = methodAndUrl.Split(' ');
+            logger.LogDebug("Checking request {request}...", methodAndUrl);
+            var (method, url) = (methodAndUrlChunks[0].ToUpper(), methodAndUrlChunks[1]);
+
+            var scopesFromTheToken = MinimalPermissionsUtils.GetScopesFromToken(request.Context?.Session.HttpClient.Request.Headers.First(h => h.Name.Equals("authorization", StringComparison.OrdinalIgnoreCase)).Value, logger);
+            if (scopesFromTheToken.Length != 0)
+            {
+                tokenPermissions.AddRange(scopesFromTheToken);
+            }
+            else
+            {
+                errors.Add(new()
+                {
+                    Request = methodAndUrl,
+                    Error = "No scopes found in the token"
+                });
+            }
+
+            // get allowed scopes for the operation
+            if (!Enum.TryParse<OperationType>(method, true, out var operationType))
+            {
+                errors.Add(new()
+                {
+                    Request = methodAndUrl,
+                    Error = $"{method} is not a valid HTTP method"
+                });
+                continue;
+            }
+
+            var pathItem = openApiDocument.FindMatchingPathItem(url, logger);
+            if (pathItem is null)
+            {
+                errors.Add(new()
+                {
+                    Request = methodAndUrl,
+                    Error = "No matching path item found"
+                });
+                continue;
+            }
+
+            if (!pathItem.Value.Value.Operations.TryGetValue(operationType, out var operation))
+            {
+                errors.Add(new()
+                {
+                    Request = methodAndUrl,
+                    Error = "No matching operation found"
+                });
+                continue;
+            }
+
+            var scopes = operation.GetEffectiveScopes(openApiDocument, logger);
+            if (scopes.Length != 0)
+            {
+                operationsAndScopes[$"{method} {pathItem.Value.Key}"] = scopes;
+            }
+
+            operationsFromRequests.Add(new()
+            {
+                Method = operationType.ToString().ToUpper(),
+                OriginalUrl = url,
+                TokenizedUrl = pathItem.Value.Key
+            });
+        }
+
+        var (minimalScopes, unmatchedOperations) = MinimalPermissionsUtils.GetMinimalScopes(
+            operationsFromRequests
+                .Select(o => $"{o.Method} {o.TokenizedUrl}")
+                .Distinct()
+                .ToArray(),
+            operationsAndScopes
+        );
+
+        var permissionsInfo = new ApiPermissionsInfo
+        {
+            TokenPermissions = tokenPermissions,
+            OperationsFromRequests = operationsFromRequests,
+            MinimalScopes = minimalScopes,
+            UnmatchedOperations = unmatchedOperations,
+            Errors = errors
+        };
+
+        return permissionsInfo;
     }
 }
